move parsing and formatting to seperate files
