+++ include/managers/nfc/ntag_t2.h
@@ -1,0 +1,34 @@
+#pragma once
+
+#include <stdbool.h>
+#include <stddef.h>
+#include <stdint.h>
+#include "pn532.h"  // for NTAG2XX_MODEL and pn532_io_handle_t
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+bool ntag_t2_read_user_memory(pn532_io_handle_t io,
+                              uint8_t **out_buf,
+                              size_t *out_len,
+                              NTAG2XX_MODEL *out_model);
+
+bool ntag_t2_find_ndef(const uint8_t *mem,
+                       size_t mem_len,
+                       size_t *msg_off,
+                       size_t *msg_len);
+
+const char *ntag_t2_model_str(NTAG2XX_MODEL m);
+
+// Convenience: Build a details string from raw Type 2 memory
+// Returns malloc'd string which the caller must free.
+char *ntag_t2_build_details_from_mem(const uint8_t *mem,
+                                     size_t mem_len,
+                                     const uint8_t *uid,
+                                     uint8_t uid_len,
+                                     NTAG2XX_MODEL model);
+
+#ifdef __cplusplus
+}
+#endif
