+++ main/managers/views/nfc_view.c
@@ -17,6 +17,7 @@
 #include "driver/i2c.h"
 #include "pn532_driver.h"
 #include "pn532_driver_i2c.h"
+#include "managers/nfc/ntag_t2.h"
 #endif
 
 // Forward declaration of this view instance for internal references
@@ -108,45 +109,6 @@ static uint16_t g_atqa = 0;
 static uint8_t g_sak = 0;
 static NTAG2XX_MODEL g_model = NTAG2XX_UNKNOWN;
 
-// URI prefix table for NDEF URI records
-static const char* ndef_uri_prefix[] = {
-    [0x00] = NULL,
-    [0x01] = "http://www.",
-    [0x02] = "https://www.",
-    [0x03] = "http://",
-    [0x04] = "https://",
-    [0x05] = "tel:",
-    [0x06] = "mailto:",
-    [0x07] = "ftp://anonymous:anonymous@",
-    [0x08] = "ftp://ftp.",
-    [0x09] = "ftps://",
-    [0x0A] = "sftp://",
-    [0x0B] = "smb://",
-    [0x0C] = "nfs://",
-    [0x0D] = "ftp://",
-    [0x0E] = "dav://",
-    [0x0F] = "news:",
-    [0x10] = "telnet://",
-    [0x11] = "imap:",
-    [0x12] = "rtsp://",
-    [0x13] = "urn:",
-    [0x14] = "pop:",
-    [0x15] = "sip:",
-    [0x16] = "sips:",
-    [0x17] = "tftp:",
-    [0x18] = "btspp://",
-    [0x19] = "btl2cap://",
-    [0x1A] = "btgoep://",
-    [0x1B] = "tcpobex://",
-    [0x1C] = "irdaobex://",
-    [0x1D] = "file://",
-    [0x1E] = "urn:epc:id:",
-    [0x1F] = "urn:epc:tag:",
-    [0x20] = "urn:epc:pat:",
-    [0x21] = "urn:epc:raw:",
-    [0x22] = "urn:epc:",
-    [0x23] = "urn:nfc:",
-};
 
 typedef struct {
     char *text;      // allocated details text
@@ -173,216 +135,14 @@ static void nfc_set_details_async(void *ptr) {
     free(res);
 }
 
-// Helpers to read NTAG user memory and parse NDEF TLV/records
-static bool ntag_read_user_memory(pn532_io_handle_t io, uint8_t **out_buf, size_t *out_len, NTAG2XX_MODEL *out_model) {
-    if (!io || !out_buf || !out_len) return false;
-    *out_buf = NULL; *out_len = 0; if (out_model) *out_model = NTAG2XX_UNKNOWN;
-    uint8_t page0_3[16] = {0};
-    if (ntag2xx_read_page(io, 0, page0_3, sizeof(page0_3)) != ESP_OK) {
-        ESP_LOGW(TAG, "ntag_read_user_memory: page0_3 read failed");
-        return false;
-    }
-    // Capability Container at page 3
-    if (page0_3[12] != 0xE1) {
-        // Not a CC magic, still allow but mark unknown
-    }
-    uint8_t size_mul_8 = page0_3[14];
-    size_t data_bytes = size_mul_8 * 8; // user memory size
-    if (data_bytes == 0 || data_bytes > 1024) data_bytes = 1024; // safety cap
-    if (out_model) {
-        switch (size_mul_8) {
-            case 0x12: *out_model = NTAG2XX_NTAG213; break;
-            case 0x3E: *out_model = NTAG2XX_NTAG215; break;
-            case 0x6D: *out_model = NTAG2XX_NTAG216; break;
-            default: *out_model = NTAG2XX_UNKNOWN; break;
-        }
-    }
-    uint8_t *buf = (uint8_t *)malloc(data_bytes);
-    if (!buf) return false;
-    size_t copied = 0;
-    // Read starting at page 4
-    while (copied < data_bytes) {
-        if (nfc_scan_cancel) { ESP_LOGI(TAG, "ntag_read_user_memory: cancelled at start of loop (copied=%u)", (unsigned)copied); free(buf); return false; }
-        uint8_t page = 4 + (copied / 4);
-        uint8_t tmp[16] = {0};
-        if (ntag2xx_read_page(io, page, tmp, sizeof(tmp)) != ESP_OK) {
-            ESP_LOGW(TAG, "ntag_read_user_memory: read page %u failed", page);
-            free(buf);
-            return false;
-        }
-        size_t remain = data_bytes - copied;
-        size_t to_copy = (remain < sizeof(tmp)) ? remain : sizeof(tmp);
-        memcpy(buf + copied, tmp, to_copy);
-        copied += to_copy;
-        // Stop if we see TLV Terminator 0xFE and we are past it
-        for (size_t i = 0; i + 1 < to_copy; ++i) {
-            if (tmp[i] == 0xFE) { *out_buf = buf; *out_len = copied - (to_copy - i - 1); return true; }
-        }
-    }
-    *out_buf = buf;
-    *out_len = copied;
-    return true;
-}
 
-static const char* ntag_model_str(NTAG2XX_MODEL m) {
-    switch(m) {
-        case NTAG2XX_NTAG213: return "NTAG213";
-        case NTAG2XX_NTAG215: return "NTAG215";
-        case NTAG2XX_NTAG216: return "NTAG216";
-        default: return "NTAG2xx";
-    }
-}
-
-static size_t append_str(char **p, size_t *cap, const char *s) {
-    size_t l = strlen(s);
-    if (l + 1 > *cap) return 0;
-    memcpy(*p, s, l); *p += l; *cap -= l; **p = '\0'; return l;
-}
-
-static size_t append_fmt(char **p, size_t *cap, const char *fmt, ...) {
-    va_list ap; va_start(ap, fmt);
-    int n = vsnprintf(*p, *cap, fmt, ap);
-    va_end(ap);
-    if (n <= 0) return 0;
-    size_t used = (size_t)n;
-    if (used >= *cap) { used = *cap ? *cap - 1 : 0; }
-    *p += used; *cap -= used; return used;
-}
-
-// Parse NDEF payloads (URI, Text, SmartPoster subset)
-static void parse_ndef_record(uint8_t tnf, const uint8_t *type, uint8_t type_len, const uint8_t *payload, size_t payload_len, char **out, size_t *cap) {
-    if (tnf == 0x01 && type_len == 1 && type[0] == 'U' && payload_len >= 1) {
-        uint8_t code = payload[0];
-        const char *pre = (code < (sizeof(ndef_uri_prefix)/sizeof(ndef_uri_prefix[0])) ? ndef_uri_prefix[code] : NULL);
-        if (pre && strcmp(pre, "tel:") == 0) append_str(out, cap, "Phone: ");
-        else if (pre && strcmp(pre, "mailto:") == 0) append_str(out, cap, "Mail: ");
-        else append_str(out, cap, "URL: ");
-        if (pre) append_str(out, cap, pre);
-        // Rest of payload as-is
-        for (size_t i = 1; i < payload_len && *cap > 1; ++i) {
-            unsigned char c = payload[i];
-            if (c >= 32 && c <= 126) { **out = (char)c; (*out)++; (*cap)--; }
-            else { append_fmt(out, cap, "%%%02X", c); }
-        }
-        append_str(out, cap, "\n");
-        return;
-    }
-    if (tnf == 0x01 && type_len == 1 && type[0] == 'T' && payload_len >= 1) {
-        uint8_t status = payload[0];
-        uint8_t lang_len = status & 0x3F;
-        size_t text_off = 1 + lang_len;
-        if (text_off > payload_len) text_off = payload_len;
-        append_str(out, cap, "Text: ");
-        for (size_t i = text_off; i < payload_len && *cap > 1; ++i) {
-            unsigned char c = payload[i];
-            if (c >= 32 && c <= 126) { **out = (char)c; (*out)++; (*cap)--; }
-        }
-        append_str(out, cap, "\n");
-        return;
-    }
-    if (tnf == 0x01 && type_len == 2 && type[0] == 'S' && type[1] == 'p' && payload_len > 0) {
-        // SmartPoster: payload is an embedded NDEF message
-        size_t pos = 0;
-        while (pos < payload_len) {
-            if (pos + 1 > payload_len) break;
-            uint8_t flags = payload[pos++];
-            uint8_t tlen = (pos < payload_len) ? payload[pos++] : 0;
-            uint32_t plen = 0;
-            if (flags & 0x10) { // SR
-                plen = (pos < payload_len) ? payload[pos++] : 0;
-            } else {
-                if (pos + 4 > payload_len) break;
-                plen = ((uint32_t)payload[pos] << 24) | ((uint32_t)payload[pos+1] << 16) | ((uint32_t)payload[pos+2] << 8) | payload[pos+3];
-                pos += 4;
-            }
-            uint8_t idlen = (flags & 0x08) ? ((pos < payload_len) ? payload[pos++] : 0) : 0;
-            const uint8_t *tt = (pos + tlen <= payload_len) ? &payload[pos] : NULL; pos += tlen;
-            const uint8_t *id = (pos + idlen <= payload_len) ? &payload[pos] : NULL; (void)id; pos += idlen;
-            const uint8_t *pl = (pos + plen <= payload_len) ? &payload[pos] : NULL; pos += plen;
-            if (!tt || !pl) break;
-            parse_ndef_record(flags & 0x07, tt, tlen, pl, plen, out, cap);
-            if (flags & 0x40) break; // ME
-        }
-        return;
-    }
-    // Fallback: print basic info
-    append_str(out, cap, "Record: ");
-    append_fmt(out, cap, "tnf=0x%02X type=", tnf);
-    for (uint8_t i = 0; i < type_len; ++i) { if (*cap > 1) { **out = (char)type[i]; (*out)++; (*cap)--; } }
-    append_str(out, cap, "\n");
-}
-
-static char* build_ndef_details_from_t2(uint8_t *mem, size_t mem_len, const uint8_t *uid, uint8_t uid_len, NTAG2XX_MODEL model) {
-    // Find NDEF TLV 0x03
-    size_t pos = 0; size_t end = mem_len;
-    size_t msg_off = 0; size_t msg_len = 0;
-    while (pos < end) {
-        uint8_t tlv = mem[pos++];
-        if (tlv == 0x00) continue; // padding
-        if (tlv == 0xFE) break; // terminator
-        uint32_t len = 0;
-        if (pos >= end) break;
-        if (mem[pos] != 0xFF) { len = mem[pos++]; }
-        else { // 3-byte length
-            if (pos + 3 > end) break;
-            pos++;
-            len = ((uint32_t)mem[pos] << 8) | mem[pos+1];
-            pos += 2;
-        }
-        if (tlv == 0x03) { msg_off = pos; msg_len = len; break; }
-        pos += len;
-    }
-
-    size_t cap = 1024;
-    char *out = (char*)malloc(cap);
-    if (!out) return NULL;
-    char *w = out; *w = '\0';
-
-    append_fmt(&w, &cap, "Card: %s\n", ntag_model_str(model));
-    append_str(&w, &cap, "UID:");
-    for (uint8_t i = 0; i < uid_len; ++i) append_fmt(&w, &cap, " %02X", uid[i]);
-    append_str(&w, &cap, "\n");
-
-    if (msg_len == 0 || msg_off + msg_len > end) {
-        append_str(&w, &cap, "No NDEF message found\n");
-        return out;
-    }
-
-    append_fmt(&w, &cap, "NDEF message: %u bytes\n", (unsigned)msg_len);
-    size_t mpos = msg_off; size_t mend = msg_off + msg_len; int rec_idx = 0;
-    while (mpos < mend) {
-        if (mpos + 1 > mend) break;
-        uint8_t flags = mem[mpos++];
-        uint8_t tlen = (mpos < mend) ? mem[mpos++] : 0;
-        uint32_t plen = 0;
-        if (flags & 0x10) { // SR
-            plen = (mpos < mend) ? mem[mpos++] : 0;
-        } else {
-            if (mpos + 4 > mend) break;
-            plen = ((uint32_t)mem[mpos] << 24) | ((uint32_t)mem[mpos+1] << 16) | ((uint32_t)mem[mpos+2] << 8) | mem[mpos+3];
-            mpos += 4;
-        }
-        uint8_t idlen = (flags & 0x08) ? ((mpos < mend) ? mem[mpos++] : 0) : 0;
-        const uint8_t *type = (mpos + tlen <= mend) ? &mem[mpos] : NULL; mpos += tlen;
-        const uint8_t *id = (mpos + idlen <= mend) ? &mem[mpos] : NULL; (void)id; mpos += idlen;
-        const uint8_t *pl = (mpos + plen <= mend) ? &mem[mpos] : NULL; mpos += plen;
-        if (!type || !pl) break;
-        append_fmt(&w, &cap, "R%d: ", ++rec_idx);
-        parse_ndef_record(flags & 0x07, type, tlen, pl, plen, &w, &cap);
-        if (flags & 0x40) break; // ME
-    }
-    return out;
-}
 
 static void nfc_build_and_set_details(pn532_io_handle_t io, const uint8_t *uid, uint8_t uid_len) {
     uint8_t *mem = NULL; size_t mem_len = 0; NTAG2XX_MODEL model = NTAG2XX_UNKNOWN;
-    if (!ntag_read_user_memory(io, &mem, &mem_len, &model)) {
-        // Still provide basic details
+    if (!ntag_t2_read_user_memory(io, &mem, &mem_len, &model)) {
         size_t cap = 256;
         ndef_details_result_t *res = (ndef_details_result_t*)malloc(sizeof(*res));
-        if (!res) {
-            return;
-        }
+        if (!res) return;
         res->text = (char*)malloc(cap);
         res->text_len = cap;
         if (!res->text) { free(res); return; }
@@ -392,7 +152,7 @@ static void nfc_build_and_set_details(pn532_io_handle_t io, const uint8_t *uid, 
         lv_async_call(nfc_set_details_async, res);
         return;
     }
-    char *text = build_ndef_details_from_t2(mem, mem_len, uid, uid_len, model);
+    char *text = ntag_t2_build_details_from_mem(mem, mem_len, uid, uid_len, model);
     free(mem);
     if (!text) return;
     g_model = model;
@@ -856,7 +616,7 @@ static void write_flipper_nfc_file(void) {
         up += snprintf(uid_part + up, sizeof(uid_part) - up, "%02X", g_uid[i]);
         if (i + 1 < g_uid_len) up += snprintf(uid_part + up, sizeof(uid_part) - up, "-");
     }
-    const char *model_str = ntag_model_str(g_model);
+    const char *model_str = ntag_t2_model_str(g_model);
     char path[192];
     snprintf(path, sizeof(path), "%s/%s_%s.nfc", dir, model_str, uid_part);
